fix: fix importing issue for participantsContainer and timestamp
