+++ src/components/timestamp/index.ts
