+++ src/components/types.ts
@@ -42,23 +42,6 @@ export interface WebSocketClient {
   onReceive?: (handler: (event: MessageEvent) => void) => void
 }
 
-export enum ParticipantRole {
-  Owner = 'owner',
-  Member = 'member',
-}
-
-export enum ParticipantType {
-  Human = 'human',
-  Agent = 'agent',
-}
-
-export interface Participant {
-  id: string
-  displayName: string
-  participantRole: ParticipantRole
-  participantType: ParticipantType
-}
-
 export interface DataFormat {
   /** Optional title. */
   title?: string
